refactor segwit signer, tx_signature be the agnostic layer
