+++ packages/jellyfish-transaction-signature/src/index.ts
@@ -1,1 +1,2 @@
 export * from './tx_signature'
+export * from './segwit_signature'
